changes for stock postings

+++ digitz_erp/buying/doctype/purchase_invoice/purchase_invoice.py
@@ -29,15 +29,21 @@ class PurchaseInvoice(Document):
 	def add_stock_for_purchase_receipt(self):
 
 		more_records = 0
-
-		for docitem in self.items:			
-			
+		posting_date_time = get_datetime(str(self.posting_date) + " " + str(self.posting_time))
+  
+		for docitem in self.items:						
+   
+   		# For back dated purchase invoice
+			more_records_count_for_item = frappe.db.count('Stock Ledger',{'item':docitem.item,
+				'warehouse':docitem.warehouse, 'posting_date':['>', posting_date_time]})
+   
+			more_records = more_records + more_records_count_for_item
+   
+   
 			new_balance_qty = docitem.qty_in_base_unit
 			# Default valuation rate
 			valuation_rate = docitem.rate_in_base_unit						
 			
-			posting_date_time = get_datetime(str(self.posting_date) + " " + str(self.posting_time))
-
 			# Default balance value calculating withe the current row only
 			new_balance_value = new_balance_qty * valuation_rate
 
@@ -52,13 +58,16 @@ class PurchaseInvoice(Document):
 				# Find out the balance value and valuation rate. Here recalculates the total balance value and valuation rate
 				# from the balance qty in the existing rows x actual incoming rate
 
-				last_stock_ledger = frappe.db.get_value('Stock Ledger', {'item_code': ['=', docitem.item_code], 'warehouse':['=', docitem.warehouse], 'posting_date':['<', posting_date_time]},['balance_qty', 'balance_value', 'valuation_rate'],order_by='posting_date desc', as_dict=True)				
-				new_balance_qty = new_balance_qty + last_stock_ledger.balance_qty			
+				last_stock_ledger = frappe.db.get_value('Stock Ledger', {'item_code': ['=', docitem.item_code], 'warehouse':['=', docitem.warehouse], 
+                                                        'posting_date':['<', posting_date_time]},
+                                            ['balance_qty', 'balance_value', 'valuation_rate'],order_by='posting_date desc', as_dict=True)				
+    
+				if(last_stock_ledger):
+					new_balance_qty = new_balance_qty + last_stock_ledger.balance_qty			
 				
-				new_balance_value = new_balance_value + (last_stock_ledger.balance_value)
+					new_balance_value = new_balance_value + (last_stock_ledger.balance_value)
 
-				valuation_rate = new_balance_value/new_balance_qty			
-
+					valuation_rate = new_balance_value/new_balance_qty			
 		
 			new_stock_ledger = frappe.new_doc("Stock Ledger")
 			new_stock_ledger.item = docitem.item
@@ -76,37 +85,38 @@ class PurchaseInvoice(Document):
 			new_stock_ledger.source = "Purchase Invoice Item"
 			new_stock_ledger.source_document_id = docitem.name
 			new_stock_ledger.insert()	
-
-			# Find out subsequent records to recalculate and update the balances
-			if frappe.db.exists('Stock Balance', {'item':docitem.item,'warehouse': docitem.warehouse}):    
-				frappe.db.delete('Stock Balance',{'item': docitem.item, 'warehouse': docitem.warehouse} )
-
-			new_stock_balance = frappe.new_doc('Stock Balance')	
-			new_stock_balance.item = docitem.item
-			new_stock_balance.unit = docitem.unit
-			new_stock_balance.warehouse = docitem.warehouse
-			new_stock_balance.stock_qty = new_balance_qty
-			new_stock_balance.stock_value = new_balance_value
-			new_stock_balance.valuation_rate = valuation_rate
 			
-			new_stock_balance.insert()
+			# If no more records for the item, update balances. otherwise it updates in the su
+			if more_records_count_for_item==0: 
+				if frappe.db.exists('Stock Balance', {'item':docitem.item,'warehouse': docitem.warehouse}):    
+					frappe.db.delete('Stock Balance',{'item': docitem.item, 'warehouse': docitem.warehouse} )
 
-
-			item = frappe.get_doc('Item', docitem.item)
-			item.stock_balance = item.stock_balance + docitem.qty_in_base_unit
-			item.save()	
-
-			# For back dated purchase invoice
-			more_records_count = frappe.db.count('Stock Ledger',{'item':docitem.item,
-				'warehouse':docitem.warehouse, 'posting_date':['>', posting_date_time]})
+				new_stock_balance = frappe.new_doc('Stock Balance')	
+				new_stock_balance.item = docitem.item
+				new_stock_balance.unit = docitem.unit
+				new_stock_balance.warehouse = docitem.warehouse
+				new_stock_balance.stock_qty = new_balance_qty
+				new_stock_balance.stock_value = new_balance_value
+				new_stock_balance.valuation_rate = valuation_rate
 			
-			more_records = more_records + more_records_count
-		
+				new_stock_balance.insert()
+				
+				item_name = frappe.get_value("Item", docitem.item,['item_name'])
+				self.update_item_stock_balance(item_name)
+   
+			# item_balances_in_warehouses = frappe.get_list('Stock Balance',{'item': docitem.item},['stock_qty'])
+			# balance_stock_qty = 0
+			# balance_stock_value = 0
+			# for item in item_balances_in_warehouses:
+			# 	balance_stock_qty = balance_stock_qty + item.stock_qty
+			# 	balance_stock_value = balance_stock_value + item.stock_value   
+   
 		if(more_records>0):
 			self.recalculate_stock_ledgers("Add",new_stock_ledger.name)		
 
 	def on_cancel(self):	
-     
+        
+		print("after cancel event")
 		self.validate_and_cancel_purchase()
 
 		frappe.db.delete("Stock Ledger",
@@ -120,116 +130,163 @@ class PurchaseInvoice(Document):
 				})
 		
 	def validate_and_cancel_purchase(self):
-     
-		print("From validate and cancel purchase")
-		
+     	
 		posting_date_time = get_datetime(str(self.posting_date) + " " + str(self.posting_time))					
 
 		stock_ledger_count_for_more = 0
   
+		# Iterate on each item from the cancelling purchase invoice
 		for docitem in self.items:	
-			more = frappe.db.count('Stock Ledger',{'item':docitem.item,
+			more_for_item = frappe.db.count('Stock Ledger',{'item':docitem.item,
             	'warehouse':docitem.warehouse, 'posting_date':['>', posting_date_time]})
 		
-			stock_ledger_count_for_more = stock_ledger_count_for_more + more
-			subsequent_reconciliation_exists = False
+			print(more_for_item)
    
+			stock_ledger_count_for_more = stock_ledger_count_for_more + more_for_item
+   
+			# If any items in the collection has more records
 			if(stock_ledger_count_for_more>0):
+				
+				stock_ledger_items = frappe.get_list('Stock Ledger',{'item':docitem.item,
+				'warehouse':docitem.warehouse, 'posting_date':['>', posting_date_time]}, ['name','qty_in','qty_out','voucher','balance_qty','voucher_no'],order_by='posting_date')
 
-				for docitem in self.items:	
-					stock_ledger_items = frappe.get_list('Stock Ledger',{'item':docitem.item,
-					'warehouse':docitem.warehouse, 'posting_date':['>', posting_date_time]}, ['name','qty_in','qty_out','voucher','balance_qty','voucher_no'],order_by='posting_date')
+				if(stock_ledger_items):
 
-					if(stock_ledger_items):
+					qty_cancelled = docitem.qty_in_base_unit
 
-						qty_cancelled = docitem.qty_in_base_unit
+					for sl in stock_ledger_items:					
+						# On each line if outgoing qty + balance_qty (qty before outgonig) is more than the cancelling qty
+						if(sl.qty_out>0 and qty_cancelled> sl.qty_out+ sl.balance_qty):
+							frappe.throw("Cancelling the purchase is prevented due to sufficiant quantity not available for " + docitem.item +
+							" for " + sl.voucher_no)        	
+					
+		# If any items has subnsequent entries
+		# if(stock_ledger_count_for_more>0):
+		# Unconditionally do recalculation to update the balances in stock_balance and item
+		self.recalculate_stock_ledgers("Cancel","")
+    
+	def update_item_stock_balance(self, item):
+     
+		print(item)
 
-						for sl in stock_ledger_items:					
-							# Check if the record is a manual stock entry
-							if sl.voucher == "Stock Reconciliation":
-								subsequent_reconciliation_exists = True
-								break
-								
-							if(sl.qty_out>0 and qty_cancelled> sl.qty_out+ sl.balance_qty):
-								frappe.throw("Cancelling the purchase is prevented due to sufficiant quantity not available for " + docitem.item +
-								" for " + sl.voucher_no)
-        	
-			if(not subsequent_reconciliation_exists):
-				print("No subsequent_reconciliation_exists")
-				stock_balance = frappe.get_doc('Stock Balance',{'item': docitem.item, 'warehouse':docitem.warehouse})
-				stock_balance.stock_qty =  stock_balance.stock_qty - docitem.qty_in_base_unit
-				value = docitem.qty_in_base_unit * docitem.rate_in_base_unit
-				stock_balance.stock_value = stock_balance.stock_value  - value	
-				stock_balance.save()    
+		item_balances_in_warehouses = frappe.get_list('Stock Balance',{'item': item},['stock_qty','stock_value'])
+  
+		balance_stock_qty = 0
+		balance_stock_value = 0
 
-				self.recalculate_stock_ledgers("Cancel","")
+		if(item_balances_in_warehouses):
+			for item_stock in item_balances_in_warehouses:
+       
+				if item_stock.stock_qty:
+					balance_stock_qty = balance_stock_qty + item_stock.stock_qty
+
+				if item_stock.stock_value:
+					balance_stock_value = balance_stock_value + item_stock.stock_value
+	
+		item_to_update = frappe.get_doc('Item', item)	
+			
+		if(not item_to_update.stock_balance):	
+			print("not item.stock_balance")
+			item_to_update.stock_balance = 0
+			item_to_update.stock_value = 0
+		item_to_update.stock_balance = balance_stock_qty
+		item_to_update.stock_value = balance_stock_value
+			
+		item_to_update.save()	
 	
 	def recalculate_stock_ledgers(self, operation, new_stock_ledger_name):
 
         # Insert record to 'Stock Recalculate Voucher' doc
 		stock_recalc_voucher = frappe.new_doc('Stock Recalculate Voucher')
-		stock_recalc_voucher.voucher = 'Delivery Note'
+		stock_recalc_voucher.voucher = 'Purchase Invoice'
 		stock_recalc_voucher.voucher_no = self.name
 		stock_recalc_voucher.voucher_date = self.posting_date
 		stock_recalc_voucher.voucher_time = self.posting_time
 		stock_recalc_voucher.status = 'Not Started'
 		stock_recalc_voucher.source_action = operation
 		stock_recalc_voucher.insert()
-
+		
 		voucher_name = stock_recalc_voucher.name
+  
+		posting_date_time = get_datetime(str(self.posting_date) + " " + str(self.posting_time))	
+  
+		# Stock Recalculate Voucher holds the entry, and records to help the calculation in the subsequent stock ledgers
 		stock_recalc_voucher = frappe.get_doc('Stock Recalculate Voucher', voucher_name)
 		
 		for docitem in self.items:
+      
+			# more_for_item = frappe.db.count('Stock Ledger',{'item':docitem.item,
+            # 	'warehouse':docitem.warehouse, 'posting_date':['>', posting_date_time]})
+
+			# # Do not continue if more items is nto exists for the item
+			# if(more_for_item ==0):
+			# 	continue   
+   
 			balance_qty = 0;
 			valuation_rate = 0
 			balance_value = 0
+			qty_in = 0
+			qty_out =0
+			rate_in = 0
+			rate_out = 0
 			
-			if(operation == "Cancel"):			
+			if(operation == "Cancel"):	
+				
+				# For cancel purchase invoice , fetch the previous stock ledger balance. since the stock ledger corresponding to the 
+    			# cancelling invoice is not valid
 				posting_date_time = get_datetime(str(stock_recalc_voucher.voucher_date) + " " + str(stock_recalc_voucher.voucher_time))
 				previous_ledger_name = frappe.db.get_value('Stock Ledger', {'item': ['=', docitem.item], 'warehouse':['=', docitem.warehouse]
 					, 'posting_date':['<', posting_date_time]},['name'], order_by='posting_date desc', as_dict=True)
 					
-				if(previous_ledger_name):
+				if(previous_ledger_name):					
 					previous_stock_ledger = frappe.get_doc('Stock Ledger',previous_ledger_name.name)
+					qty_in = previous_stock_ledger.qty_in
+					qty_out = previous_stock_ledger.qty_out
+					rate_in = previous_stock_ledger.incomig_rate
+					rate_out = previous_stock_ledger.outgoing_rate 
 					balance_qty = previous_stock_ledger.balance_qty
 					balance_value = previous_stock_ledger.balance_value
 					valuation_rate = balance_value / balance_qty
 
 				stock_recalc_voucher.append('records',{'item': docitem.item, 
 															'warehouse': docitem.warehouse,                                                        
-															'qty_out': docitem.qty_in_base_unit,                                                        
-															'outgoing_rate':docitem.rate_in_base_unit,
+															'qty_in': qty_in,      
+															'incoming_rate': rate_in,
+															'qty_out': qty_out,
+															'outgoing_rate':rate_out,
 															'balance_qty':balance_qty,
 															'balance_value':balance_value,
 															'valuation_rate':valuation_rate,
 															'unit': docitem.unit
 						 									})
                     
-			if(operation == "Add"):	
-
-				stock_ledger = frappe.get_doc("Stock Ledger", new_stock_ledger_name)
-
+			if(operation == "Add"):
+    			
+       			# When adding purchase invoice, take the corresponding stock ledger values(newly added stock ledger in the parent method)
+				stock_ledger = frappe.get_doc("Stock Ledger", new_stock_ledger_name)				
 				stock_recalc_voucher.append('records',{'item': docitem.item, 
-															'warehouse': docitem.warehouse,                                                        
-															'qty_out': docitem.qty_in_base_unit,                                                        
-															'outgoing_rate':docitem.rate_in_base_unit,
-															'balance_qty':stock_ledger.balance_qty,
-															'balance_value':stock_ledger.balance_value,
-															'valuation_rate':stock_ledger.valuation_rate,
-															'unit': docitem.unit
-															})
-
-			stock_recalc_voucher.save()
+														'warehouse': docitem.warehouse,                                                        
+														'qty_in': stock_ledger.qty_in,      
+														'incoming_rate': stock_ledger.incoming_rate,
+														'qty_out': stock_ledger.qty_out,
+														'outgoing_rate':stock_ledger.outgoing_rate,
+														'balance_qty':balance_qty,
+														'balance_value':balance_value,
+														'valuation_rate':valuation_rate,
+														'unit': docitem.unit
+														})
+		# stock_recalc_voucher.save()
+		self.recalculate_subsequent_stock_ledgers(stock_recalc_voucher)
          
-			self.recalculate_subsequent_stock_ledgers(stock_recalc_voucher.name)
-         
-			frappe.msgprint("Stock balances updated successfully.")
+			# frappe.msgprint("Stock balances updated successfully.")
 	
-	def recalculate_subsequent_stock_ledgers(self,stock_recalc_voucher_name):
-
-		stock_recalc_voucher = frappe.get_doc("Stock Recalculate Voucher", stock_recalc_voucher_name)      
+	def recalculate_subsequent_stock_ledgers(self,stock_recalc_voucher):
+     
+		# stock_recalc_voucher = frappe.get_doc("Stock Recalculate Voucher", stock_recalc_voucher_name)      
 		stock_recalc_voucher.status = 'Started'
 		stock_recalc_voucher.start_time = now()
+		print(stock_recalc_voucher)
+		print(stock_recalc_voucher.records)
 
 		posting_date_time = get_datetime(str(stock_recalc_voucher.voucher_date) + " " + str(stock_recalc_voucher.voucher_time))			
 
@@ -238,27 +295,52 @@ class PurchaseInvoice(Document):
 			balance_qty = record.balance_qty
 			balance_value = record.balance_value
 			valuation_rate = record.valuation_rate
+			qty_in = record.qty_in
+			incoming_rate = record.incoming_rate
+			qty_out = record.qty_out
+			outgoing_rate = record.outgoing_rate			
+			
 
 			stock_ledger_items = frappe.get_list('Stock Ledger',{'item':record.item,
 			'warehouse':record.warehouse, 'posting_date':['>', posting_date_time]}, 'name',order_by='posting_date')
+   
+			# Scenario 1- PUrchase Invoice - current row cancelled. For this assigned previous stock ledger balance to 'Stock Recalculate Voucher'
 
 			for sl_name in stock_ledger_items:
 				sl = frappe.get_doc('Stock Ledger', sl_name)
+				print(sl)
 				
 				if(sl.voucher == "Delivery Note"):
-					balance_qty = balance_qty - record.qty_out
+					balance_qty = balance_qty - sl.qty_out
 					balance_value = balance_qty * valuation_rate
 				
 				if (sl.voucher == "Purchase Invoice"):
-					balance_qty = balance_qty + record.qty_in
-					balance_value = balance_value  + (record.qty_in * record.incoming_rate)
-					valuation_rate = balance_value/ balance_qty
+					balance_qty = balance_qty + sl.qty_in
+					balance_value = balance_value  + (sl.qty_in * sl.incoming_rate)
+					if(balance_qty!=0): #Avoid divisible by zero
+						valuation_rate = balance_value/ balance_qty						
 				
 				sl.balance_qty = balance_qty
 				sl.balance_value = balance_value
 				sl.valuation_rate = valuation_rate
 				
 				sl.save()
+    
+			if frappe.db.exists('Stock Balance', {'item':record.item,'warehouse': record.warehouse}):    
+				frappe.db.delete('Stock Balance',{'item': record.item, 'warehouse': record.warehouse} )
+
+			new_stock_balance = frappe.new_doc('Stock Balance')	
+			new_stock_balance.item = record.item
+			new_stock_balance.unit = record.unit
+			new_stock_balance.warehouse = record.warehouse
+			new_stock_balance.stock_qty = balance_qty
+			new_stock_balance.stock_value = balance_value
+			new_stock_balance.valuation_rate = valuation_rate
+
+			new_stock_balance.insert()
+
+			item_name = frappe.get_value("Item", record.item,['item_name'])
+			self.update_item_stock_balance(item_name)					
 
 		stock_recalc_voucher.status = 'Completed'
 		stock_recalc_voucher.end_time = now()
